[DF] Split column reader's Get into Load+Get
This is in preparation of bulk entry processing in RDF.
Rather than having a single method to fetch an entry value and
return it, column readers now do this in two steps:
the `Load` method is responsible of loading values and
needs to know the actual entry numbers and their mask.
The `Get` method only needs an index into the array of
values for the current bulk.

Currently entry and mask are scalars. Future commits will
work towards switching them to to arrays, actually enabling
bulk entry processing.