[DF] Bulkify RTreeColumnReader, loop over TTree entries in bulk
This commit teaches RTreeColumnReader how to work with bulk of
branch values and RLoopManager how to loop over TTree entries
in bulk. It supports the single-thread and multi-thread case.

Since we are now always copying branch values into our own cache,
this breaks a number of features:
  - accessing column values of type `T` as `BaseOfT&` (e.g. using
    TH1& to access a TH1F column value, or using TObject& to read
    a branch with polymorphic values)
  - working with columns of non-copiable types (e.g. unique_ptr<T>)
  - working with columns of types lacking a default constructor
    (even if they have an I/O constructor)
  - reading TClonesArrays now breaks at runtime because it is not
    possible to copy-assign a TClonesArray to a default-constructed
    TClonesArray (because of a fClass equality check in
    TClonesArray::operator=)

Tests that would now fail to compile have been disabled in this
commit.

Notably we don't need to specialize RTreeColumnReader for the
case of boolean arrays anymore because we always copy array
contents independently of whether the I/O layer provides
contiguous storage or not.