[VecOps] Add VecOps::ResetView
This is a low-level function to reset the buffer for which this RVec
is acting as a view. Before this commit, the only way to reset
the pointed-to address of a non-owning RVec was to construct a new
RVec and swap it with the old one, which is much more expensive
than the two assignments ResetView performs.